Update solargame.js
Sooo I the array to keep track of how many planets have been visited is called numberOfPlanets. the other functions are checkPlanetNumber()---if the length of the array=8 then you can go to the sun, if not then you can't go---and addPlanet() adds another element to the array for every planet you go to (called when you run the Doom function). if you type in 'sun' to the travel page, no matter what you have done already, the page doesn't change or anything (should change to page 11, the new page a made for the sun) and gives us no error.